# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/menu": {
      "get": {
        "description": "Возвращает список блюд меню с поддержкой:\n    - Поиска по названию\n    - Фильтрации по категории, типу блюда, цене, времени приготовления, калориям\n    - Сортировки по различным полям\n    - Пагинации\n    Показываются только неудаленные блюда (is_deleted = false)",
        "operationId": "MenuController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Поиск по названию блюда",
            "schema": {
              "example": "хачапури",
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "description": "Фильтр по категории",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "is_vegetarian",
            "required": false,
            "in": "query",
            "description": "Фильтр только вегетарианских блюд",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "is_spicy",
            "required": false,
            "in": "query",
            "description": "Фильтр только острых блюд",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "min_price",
            "required": false,
            "in": "query",
            "description": "Минимальная цена",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "max_price",
            "required": false,
            "in": "query",
            "description": "Максимальная цена",
            "schema": {
              "example": 100,
              "type": "number"
            }
          },
          {
            "name": "max_cooking_time",
            "required": false,
            "in": "query",
            "description": "Максимальное время приготовления",
            "schema": {
              "example": 30,
              "type": "number"
            }
          },
          {
            "name": "max_calories",
            "required": false,
            "in": "query",
            "description": "Максимальное количество калорий",
            "schema": {
              "example": 500,
              "type": "number"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "description": "Поле для сортировки",
            "schema": {
              "example": "price",
              "type": "string",
              "enum": [
                "price",
                "cooking_time_minutes",
                "calories",
                "item_name",
                "created_at"
              ]
            }
          },
          {
            "name": "sort_order",
            "required": false,
            "in": "query",
            "description": "Порядок сортировки",
            "schema": {
              "example": "ASC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Номер страницы (для пагинации)",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Количество элементов на странице",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список блюд успешно получен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuPaginationResponseDto"
                }
              }
            }
          }
        },
        "summary": "Получить меню с фильтрацией и поиском",
        "tags": [
          "Menu"
        ]
      },
      "post": {
        "description": "Создает новое блюдо в меню",
        "operationId": "MenuController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Блюдо успешно создано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные для создания блюда или категория не найдена"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Создать новое блюдо",
        "tags": [
          "Menu"
        ]
      }
    },
    "/menu/category/{categoryId}": {
      "get": {
        "description": "Возвращает список блюд определенной категории с поддержкой фильтрации",
        "operationId": "MenuController_findByCategory",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "description": "ID категории меню",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Поиск по названию блюда",
            "schema": {
              "example": "хачапури",
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "description": "Фильтр по категории",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "is_vegetarian",
            "required": false,
            "in": "query",
            "description": "Фильтр только вегетарианских блюд",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "is_spicy",
            "required": false,
            "in": "query",
            "description": "Фильтр только острых блюд",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "min_price",
            "required": false,
            "in": "query",
            "description": "Минимальная цена",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "max_price",
            "required": false,
            "in": "query",
            "description": "Максимальная цена",
            "schema": {
              "example": 100,
              "type": "number"
            }
          },
          {
            "name": "max_cooking_time",
            "required": false,
            "in": "query",
            "description": "Максимальное время приготовления",
            "schema": {
              "example": 30,
              "type": "number"
            }
          },
          {
            "name": "max_calories",
            "required": false,
            "in": "query",
            "description": "Максимальное количество калорий",
            "schema": {
              "example": 500,
              "type": "number"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "description": "Поле для сортировки",
            "schema": {
              "example": "price",
              "type": "string",
              "enum": [
                "price",
                "cooking_time_minutes",
                "calories",
                "item_name",
                "created_at"
              ]
            }
          },
          {
            "name": "sort_order",
            "required": false,
            "in": "query",
            "description": "Порядок сортировки",
            "schema": {
              "example": "ASC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Номер страницы (для пагинации)",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Количество элементов на странице",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Блюда категории успешно получены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuPaginationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Категория с указанным ID не найдена"
          }
        },
        "summary": "Получить блюда по категории",
        "tags": [
          "Menu"
        ]
      }
    },
    "/menu/{id}": {
      "get": {
        "description": "Возвращает детальную информацию о блюде с информацией о категории",
        "operationId": "MenuController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор блюда",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Блюдо найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Блюдо с указанным ID не найдено"
          }
        },
        "summary": "Получить блюдо по ID",
        "tags": [
          "Menu"
        ]
      },
      "patch": {
        "description": "Обновляет информацию о блюде",
        "operationId": "MenuController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор блюда",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Блюдо успешно обновлено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные для обновления блюда"
          },
          "404": {
            "description": "Блюдо с указанным ID не найдено"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить блюдо",
        "tags": [
          "Menu"
        ]
      },
      "delete": {
        "description": "Полностью удаляет блюдо из базы данных",
        "operationId": "MenuController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор блюда",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Блюдо успешно удалено"
          },
          "404": {
            "description": "Блюдо с указанным ID не найдено"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить блюдо",
        "tags": [
          "Menu"
        ]
      }
    },
    "/menu/{id}/soft-delete": {
      "patch": {
        "description": "Помечает блюдо как удаленное (is_deleted = true), но не удаляет из БД",
        "operationId": "MenuController_softDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор блюда",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Блюдо успешно помечено как удаленное",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Блюдо с указанным ID не найдено"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Мягкое удаление блюда",
        "tags": [
          "Menu"
        ]
      }
    },
    "/menu/{id}/restore": {
      "patch": {
        "description": "Восстанавливает мягко удаленное блюдо (is_deleted = false)",
        "operationId": "MenuController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор блюда",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Блюдо успешно восстановлено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Блюдо с указанным ID не найдено"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Восстановить удаленное блюдо",
        "tags": [
          "Menu"
        ]
      }
    },
...
  },
  "info": {
    "title": "Georgian Restaurant API",
    "description": "API для управления грузинскими ресторанами с системой бронирования, заказов и аудита",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "MenuItemResponseDto": {
        "type": "object",
        "properties": {
          "item_id": {
            "type": "number",
            "description": "Уникальный идентификатор блюда",
            "example": 1
          },
          "item_name": {
            "type": "string",
            "description": "Название блюда",
            "example": "Хачапури по-аджарски",
            "minLength": 3,
            "maxLength": 150
          },
          "item_description": {
            "type": "string",
            "description": "Описание блюда",
            "example": "Традиционная грузинская лепешка с сыром и яйцом"
          },
          "category_id": {
            "type": "number",
            "description": "ID категории меню",
            "example": 1
          },
          "price": {
            "type": "number",
            "description": "Стоимость блюда",
            "example": 25.5
          },
          "cooking_time_minutes": {
            "type": "number",
            "description": "Время приготовления в минутах",
            "example": 15
          },
          "calories": {
            "type": "number",
            "description": "Калории",
            "example": 350
          },
          "is_vegetarian": {
            "type": "boolean",
            "description": "Вегетарианское блюдо",
            "example": false
          },
          "is_spicy": {
            "type": "boolean",
            "description": "Острое блюдо",
            "example": false
          },
          "is_deleted": {
            "type": "boolean",
            "description": "Удалено (мягкое удаление)",
            "example": false
          },
          "image_url": {
            "type": "string",
            "description": "URL изображения блюда",
            "example": "https://example.com/image.jpg"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Дата последнего обновления",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "category": {
            "type": "object",
            "description": "Информация о категории"
          }
        },
        "required": [
          "item_id",
          "item_name",
          "category_id",
          "price",
          "cooking_time_minutes",
          "is_vegetarian",
          "is_spicy",
          "is_deleted",
          "created_at",
          "updated_at"
        ]
      },
      "MenuPaginationResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Список блюд",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MenuItemResponseDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Общее количество блюд",
            "example": 45
          },
          "page": {
            "type": "number",
            "description": "Текущая страница",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Количество элементов на странице",
            "example": 20
          },
          "pages": {
            "type": "number",
            "description": "Общее количество страниц",
            "example": 3
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "pages"
        ]
      },
      "CreateMenuItemDto": {
        "type": "object",
        "properties": {
          "item_name": {
            "type": "string",
            "description": "Название блюда",
            "example": "Хачапури по-аджарски",
            "minLength": 3,
            "maxLength": 150
          },
          "item_description": {
            "type": "string",
            "description": "Описание блюда",
            "example": "Традиционная грузинская лепешка с сыром и яйцом"
          },
          "category_id": {
            "type": "number",
            "description": "ID категории меню",
            "example": 1
          },
          "price": {
            "type": "number",
            "description": "Стоимость блюда",
            "example": 25.5
          },
          "cooking_time_minutes": {
            "type": "number",
            "description": "Время приготовления в минутах",
            "example": 15,
            "default": 0
          },
          "calories": {
            "type": "number",
            "description": "Калории",
            "example": 350
          },
          "is_vegetarian": {
            "type": "boolean",
            "description": "Вегетарианское блюдо",
            "example": false,
            "default": false
          },
          "is_spicy": {
            "type": "boolean",
            "description": "Острое блюдо",
            "example": false,
            "default": false
          },
          "image_url": {
            "type": "string",
            "description": "URL изображения блюда",
            "example": "https://example.com/image.jpg"
          }
        },
        "required": [
          "item_name",
          "category_id",
          "price"
        ]
      },
      "UpdateMenuItemDto": {
        "type": "object",
        "properties": {
          "item_name": {
            "type": "string",
            "description": "Название блюда",
            "example": "Хачапури по-аджарски",
            "minLength": 3,
            "maxLength": 150
          },
          "item_description": {
            "type": "string",
            "description": "Описание блюда",
            "example": "Традиционная грузинская лепешка с сыром и яйцом"
          },
          "category_id": {
            "type": "number",
            "description": "ID категории меню",
            "example": 1
          },
          "price": {
            "type": "number",
            "description": "Стоимость блюда",
            "example": 25.5
          },
          "cooking_time_minutes": {
            "type": "number",
            "description": "Время приготовления в минутах",
            "example": 15
          },
          "calories": {
            "type": "number",
            "description": "Калории",
            "example": 350
          },
          "is_vegetarian": {
            "type": "boolean",
            "description": "Вегетарианское блюдо",
            "example": false
          },
          "is_spicy": {
            "type": "boolean",
            "description": "Острое блюдо",
            "example": false
          },
          "image_url": {
            "type": "string",
            "description": "URL изображения блюда",
            "example": "https://example.com/image.jpg"
          }
        }
      },
...
    }
  }
}